Refactoring Begin
Extract embedded PluginConfig class and move it to a separate unit
