+++ PluginConfig.cpp
@@ -1,0 +1,1 @@
+#include "PluginConfig.h"
